+++ forte/data/data_pack_dataset.py
@@ -16,9 +16,9 @@ from typing import Dict, Iterator, Type, Optional, List, Tuple, Union
 
 from forte.data.readers.base_reader import PackReader
 from forte.data.data_pack import DataPack
-from forte.data.extractor.converter import Converter
+from forte.data.converter.converter import Converter
 from forte.data.extractor.extractor import BaseExtractor
-from forte.data.extractor.feature import Feature
+from forte.data.converter.feature import Feature
 from forte.data.ontology.top import Annotation
 from forte.data.ontology.core import EntryType
 from forte.data.types import DataRequest
